# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口文档",
        "title": "go-server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/private/v1/goods": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "列表 分页",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "SortAsc",
                            "SortDesc"
                        ],
                        "description": "排序",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goodsdto.PageResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/goodsdto.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/goodsdto.DetailPublicDTO"
                        }
                    }
                }
            }
        },
        "/api/private/v1/goods/info": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/goodsdto.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goodsdto.DetailPublicDTO"
                        }
                    }
                }
            }
        },
        "/api/private/v1/goods/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "获取详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goodsdto.DetailPublicDTO"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/private/v1/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/managerdto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/managerdto.LoginResponse"
                        }
                    }
                }
            }
        },
...
        "/api/private/v1/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "管理员列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/managerdto.PageResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/managerdto.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "管理员创建",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/managerdto.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/private/v1/users/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "goodsdto.AttrDTO": {
            "type": "object",
            "properties": {
                "add_price": {
                    "description": "该属性需要额外增加的价钱",
                    "type": "number"
                },
                "attr_id": {
                    "description": "属性id",
                    "type": "integer"
                },
                "attr_value": {
                    "description": "商品对应属性的值",
                    "type": "string"
                },
                "goods_id": {
                    "description": "商品id",
                    "type": "integer"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                }
            }
        },
        "goodsdto.CreateAttr": {
            "type": "object",
            "required": [
                "attr_id",
                "attr_value"
            ],
            "properties": {
                "attr_id": {
                    "type": "integer"
                },
                "attr_value": {
                    "type": "string"
                }
            }
        },
        "goodsdto.CreatePics": {
            "type": "object",
            "required": [
                "pic",
                "url"
            ],
            "properties": {
                "pic": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "goodsdto.CreateRequest": {
            "type": "object",
            "required": [
                "attrs",
                "goods_cat",
                "goods_introduce",
                "goods_name",
                "goods_number",
                "goods_price",
                "goods_weight",
                "pics"
            ],
            "properties": {
                "attrs": {
                    "description": "商品属性",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.CreateAttr"
                    }
                },
                "goods_cat": {
                    "description": "商品分类",
                    "type": "string"
                },
                "goods_introduce": {
                    "description": "商品介绍",
                    "type": "string"
                },
                "goods_name": {
                    "description": "gorm.Model",
                    "type": "string"
                },
                "goods_number": {
                    "description": "商品数量",
                    "type": "integer"
                },
                "goods_price": {
                    "description": "商品价格",
                    "type": "number"
                },
                "goods_weight": {
                    "description": "商品重量",
                    "type": "integer"
                },
                "pics": {
                    "description": "商品图片",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.CreatePics"
                    }
                }
            }
        },
        "goodsdto.DetailPublicDTO": {
            "type": "object",
            "properties": {
                "add_time": {
                    "description": "添加时间s",
                    "type": "integer"
                },
                "attrs": {
                    "description": "商品属性",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.AttrDTO"
                    }
                },
                "cat_id": {
                    "description": "商品分类",
                    "type": "integer"
                },
                "cat_one_id": {
                    "description": "商品分类",
                    "type": "integer"
                },
                "cat_three_id": {
                    "description": "商品分类",
                    "type": "integer"
                },
                "cat_two_id": {
                    "description": "商品分类",
                    "type": "integer"
                },
                "goods_big_logo": {
                    "description": "商品大图",
                    "type": "string"
                },
                "goods_cat": {
                    "description": "商品分类",
                    "type": "string"
                },
                "goods_id": {
                    "description": "主键",
                    "type": "integer"
                },
                "goods_introduce": {
                    "description": "商品介绍",
                    "type": "string"
                },
                "goods_name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "goods_number": {
                    "description": "商品数量",
                    "type": "integer"
                },
                "goods_price": {
                    "description": "商品价格",
                    "type": "number"
                },
                "goods_small_logo": {
                    "description": "商品小图",
                    "type": "string"
                },
                "goods_state": {
                    "description": "商品状态",
                    "type": "integer"
                },
                "goods_weight": {
                    "description": "商品重量",
                    "type": "integer"
                },
                "is_promote": {
                    "description": "是否促销",
                    "type": "integer"
                },
                "pics": {
                    "description": "商品图片",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.PicDTO"
                    }
                },
                "upd_time": {
                    "description": "更新时间",
                    "type": "integer"
                }
            }
        },
        "goodsdto.PageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.PublicDTO"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "goodsdto.PicDTO": {
            "type": "object",
            "properties": {
                "goods_id": {
                    "description": "商品id",
                    "type": "integer"
                },
                "pics_big_url": {
                    "description": "相册大图",
                    "type": "string"
                },
                "pics_id": {
                    "description": "主键",
                    "type": "integer"
                },
                "pics_mid_url": {
                    "description": "相册中图",
                    "type": "string"
                },
                "pics_sma_url": {
                    "description": "相册小图s",
                    "type": "string"
                }
            }
        },
        "goodsdto.PublicDTO": {
            "type": "object",
            "properties": {
                "add_time": {
                    "description": "添加时间s",
                    "type": "integer"
                },
                "goods_id": {
                    "description": "主键",
                    "type": "integer"
                },
                "goods_name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "goods_number": {
                    "description": "商品数量",
                    "type": "integer"
                },
                "goods_price": {
                    "description": "商品价格",
                    "type": "number"
                },
                "goods_state": {
                    "description": "商品状态",
                    "type": "integer"
                },
                "goods_weight": {
                    "description": "商品重量",
                    "type": "integer"
                },
                "is_promote": {
                    "description": "是否促销",
                    "type": "integer"
                }
            }
        },
...
        "goodsdto.UpdateRequest": {
            "type": "object",
            "required": [
                "attrs",
                "goods_cat",
                "goods_introduce",
                "goods_name",
                "goods_number",
                "goods_price",
                "goods_weight",
                "pics"
            ],
            "properties": {
                "attrs": {
                    "description": "商品属性",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.CreateAttr"
                    }
                },
                "goods_cat": {
                    "description": "商品分类",
                    "type": "string"
                },
                "goods_introduce": {
                    "description": "商品介绍",
                    "type": "string"
                },
                "goods_name": {
                    "description": "gorm.Model",
                    "type": "string"
                },
                "goods_number": {
                    "description": "商品数量",
                    "type": "integer"
                },
                "goods_price": {
                    "description": "商品价格",
                    "type": "number"
                },
                "goods_weight": {
                    "description": "商品重量",
                    "type": "integer"
                },
                "pics": {
                    "description": "商品图片",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goodsdto.CreatePics"
                    }
                }
            }
        },
        "managerdto.CreateRequest": {
            "type": "object",
            "required": [
                "email",
                "mobile",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "managerdto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "managerdto.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "mg_state": {
                    "description": "状态",
                    "type": "boolean"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色id",
                    "type": "integer"
                },
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "managerdto.ManagerPublicDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "managerdto.PageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/managerdto.ManagerPublicDTO"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "managerdto.UpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "state": {
                    "description": "状态",
                    "type": "boolean"
                }
            }
        },
...
    }
}